# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Firetower",
    "description": "The Firetower control plane: API, scheduling, and worker transports.",
    "contact": {
      "name": "Westlabs LLC"
    },
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.10.0"
  },
  "paths": {
...
    "/api/v1/agents/{kind}": {
      "put": {
        "tags": [
          "agents"
        ],
        "summary": "Configure how an agent authenticates.",
        "operationId": "configure_agent",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "description": "Agent kind",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigureAgent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "agents"
        ],
        "summary": "Forget an agent's configuration and any credential with it.",
        "operationId": "forget_agent",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "description": "Agent kind",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Replay. The live feed is the event stream; this is the backfill after a hard\nrefresh, and the fallback when a stream can't be held open.",
        "operationId": "list_events",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "description": "Last sequence number seen",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "sessionId",
            "in": "query",
            "description": "Only this session's events",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Agent": {
        "type": "string",
        "description": "Which agent runs inside a workspace.\n\nSerialised as the variant name — see the wire conventions in the brief: a\nfield takes the consumer's casing, an enum value stays the symbol it is.",
        "enum": [
          "ClaudeCode",
          "Codex",
          "Shell"
        ]
      },
      "AgentMode": {
        "type": "string",
        "description": "How an agent proves who it is.",
        "enum": [
          "Subscription",
          "ApiKey",
          "NotNeeded"
        ]
      },
...
      "ApiError": {
        "type": "object",
        "description": "Every non-success response, so failures are as typed as everything else.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "description": "For humans and logs. The interface should switch on `code` and write its\nown copy — only it knows the context and what to offer next."
          }
        }
      },
...
      "ConfigureAgent": {
        "type": "object",
        "required": [
          "mode"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "mode": {
            "$ref": "#/components/schemas/AgentMode"
          },
          "secret": {
            "type": [
              "string",
              "null"
            ],
            "description": "The token from `claude setup-token`, or a metered API key — whichever\nthe mode calls for. Required for both; ignored for an agent that needs\nno credential."
          }
        }
      },
...
      "ErrorCode": {
        "type": "string",
        "description": "The catalogue is the type, so there is no separate list to keep in sync.",
        "enum": [
          "InvalidRequest",
          "NotFound",
          "NoCapacity",
          "HostUnreachable",
          "RepoNotConnected",
          "SessionEnded",
          "ProviderNotConfigured",
          "ProviderNotConnected",
          "RepoAccessDenied",
          "RepoUnreachable",
          "RepoUnusable",
          "RepoInUse",
          "ActionFailed",
          "Unauthorized",
          "PasswordChangeRequired",
          "Internal"
        ]
      },
      "Event": {
        "type": "object",
        "description": "Something that happened, recorded by the worker that it happened on.\n\nSessions are a projection of these, never the other way round.",
        "required": [
          "seq",
          "sessionId",
          "kind",
          "at"
        ],
        "properties": {
          "at": {
            "type": "string",
            "format": "date-time"
          },
          "kind": {
            "$ref": "#/components/schemas/EventKind"
          },
          "seq": {
            "type": "integer",
            "format": "int64",
            "description": "Monotonic per worker. This is the resume cursor."
          },
          "sessionId": {
            "$ref": "#/components/schemas/SessionId"
          }
        }
      },
      "EventKind": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "repo",
              "prompt",
              "type"
            ],
            "properties": {
              "prompt": {
                "type": "string"
              },
              "repo": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "SessionCreated"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "host",
              "detail",
              "type"
            ],
            "properties": {
              "detail": {
                "type": "string"
              },
              "host": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "HostSelected"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "A step is under way. Every other event here is a *completion*, which is\nwhy a session that spent eight minutes fetching a repository looked\nfrozen: nothing had happened yet, so nothing had been said.",
            "required": [
              "step",
              "type"
            ],
            "properties": {
              "step": {
                "$ref": "#/components/schemas/Step"
              },
              "type": {
                "type": "string",
                "enum": [
                  "StepStarted"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "How a step is getting on, while it is still going. Sent sparingly — one\nline replacing the last, not a log.",
            "required": [
              "step",
              "detail",
              "type"
            ],
            "properties": {
              "detail": {
                "type": "string"
              },
              "step": {
                "$ref": "#/components/schemas/Step"
              },
              "type": {
                "type": "string",
                "enum": [
                  "StepProgress"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "detail",
              "type"
            ],
            "properties": {
              "detail": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "RepoFetched"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "branch",
              "type"
            ],
            "properties": {
              "askedFor": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "The name asked for, when it is not the name granted.\n\nTwo sessions started from the same prompt want the same branch, so\nthe second is numbered. That is worth saying: the branch is what a\npull request is opened from, and finding out it was renamed by\nreading the pull request is finding out too late.\n\nRenamed explicitly: this enum carries no `rename_all`, and every\nother field in it happens to be one word, so it is the first that\nwould otherwise have reached the API as snake_case."
              },
              "branch": {
                "type": "string"
              },
              "repo": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "Which repository, when a session has more than one.\n\nAbsent from a worker that predates a session holding more than one,\nand from a session that holds exactly one — in both cases there is\nnothing to disambiguate."
              },
              "type": {
                "type": "string",
                "enum": [
                  "WorktreeAdded"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "detail",
              "type"
            ],
            "properties": {
              "detail": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "WorkspaceStarted"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "detail",
              "type"
            ],
            "properties": {
              "detail": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "SetupFinished"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "name",
              "type"
            ],
            "properties": {
              "name": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "TmuxOpened"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "agent",
              "type"
            ],
            "properties": {
              "agent": {
                "$ref": "#/components/schemas/Agent"
              },
              "type": {
                "type": "string",
                "enum": [
                  "AgentLaunched"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "status",
              "type"
            ],
            "properties": {
              "note": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "Why, when whatever changed it knows.\n\nThe agent's own words: the permission it is asking for, the last\nthing it said before finishing, the error that stopped it. Without\nthis a blocked session is a red dot you have to open a terminal to\nunderstand, which is most of the cost of being interrupted."
              },
              "status": {
                "$ref": "#/components/schemas/SessionStatus"
              },
              "type": {
                "type": "string",
                "enum": [
                  "StatusChanged"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "code",
              "message",
              "type"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Failed"
                ]
              }
            }
          }
        ]
      },
...
      "SessionId": {
        "type": "string",
        "description": "Identifies a session — the unit of work you talk to."
      },
      "SessionStatus": {
        "type": "string",
        "enum": [
          "Starting",
          "Working",
          "NeedsYou",
          "HandedBack",
          "Failed",
          "Ended"
        ]
      },
...
      "Step": {
        "type": "string",
        "description": "One stage of bringing a session up.\n\nThe point of naming them is that the whole list is knowable *before* any of\nit runs — so a session can show what it is going to do the moment it is\ncreated, rather than assembling a shape out of events as they arrive. A step\nnobody has reached yet is still worth showing.",
        "enum": [
          "Fetch",
          "Worktree",
          "Workspace",
          "Setup",
          "Launch"
        ]
      },
...
    }
  }
}